[maeve_geometry] fix wrong typedef

+++ maeve_geometry/include/maeve_automation_core/maeve_geometry/interval.h
@@ -406,7 +406,7 @@ class Interval {
 //------------------------------------------------------------------------------
 
 typedef Interval<double> Interval_d;
-typedef Interval<double> Interval_f;
+typedef Interval<float> Interval_f;
 
 //------------------------------------------------------------------------------
